requirements_test.txt
django-cms==3.4.5
git+https://github.com/divio/djangocms-attributes-field
flake8==3.0.4
